Tweak programs

+++ mfiller/src/main.rs
@@ -1,3 +1,4 @@
+use byte_unit::{Byte, ByteError};
 use clap::Parser;
 use rand::seq::SliceRandom;
 use std::fmt;
@@ -7,10 +8,10 @@ use std::path::{Path, PathBuf};
 
 #[derive(Parser, Debug)]
 struct CliArgs {
-    #[clap(parse(try_from_str = parse_size_string))]
-    file_size: u64,
+    #[clap(short='s', long, default_value = "1M", parse(try_from_str = parse_size_string))]
+    file_size: u128,
 
-    #[clap()]
+    #[clap(short = 'c', long, default_value_t = 300)]
     file_count: u16,
 }
 
@@ -23,46 +24,39 @@ impl fmt::Display for ParseError {
 }
 impl std::error::Error for ParseError {}
 
-fn parse_size_string(size: &str) -> Result<u64, ParseError> {
-    let x = size.to_uppercase();
-    let multiplier = match x.chars().last().ok_or(ParseError {})? {
-        'K' => 1 << 10,
-        'M' => 1 << 20,
-        'G' => 1 << 30,
-        _ => 1,
-    };
-    let integer_value = x[..x.len() - 1].parse::<u64>().map_err(|_| ParseError {})?;
-    Ok(integer_value * multiplier)
+fn parse_size_string(size: &str) -> Result<u128, ByteError> {
+    Ok(Byte::from_str(size)?.get_bytes())
 }
 
-fn write_content(file_path: &PathBuf, n: u64) {
+fn write_content(file_path: &PathBuf, n: u128) {
     let from_file = File::open("/dev/random").expect("Could not open /dev/random");
     let mut reader = BufReader::new(from_file);
     let to_file = OpenOptions::new()
+        .create(true)
         .write(true)
         .open(file_path)
         .expect("Could not open file for write");
     let mut writer = BufWriter::new(to_file);
     let mut buf = [0; 1 << 16];
 
-    let mut bytes_copied: u64 = 0;
+    let mut bytes_copied: u128 = 0;
     while bytes_copied < n {
-        let bytes_to_read = std::cmp::min(buf.len() as u64, n - bytes_copied);
-        let bytes_read = reader
+        let bytes_to_read = std::cmp::min(buf.len() as u128, n - bytes_copied);
+        let bytes_read: u128 = reader
             .read(&mut buf[..bytes_to_read as usize])
-            .expect("Could not read");
+            .expect("Could not read") as u128;
         if bytes_read == 0 {
             break;
         }
-        let bytes_to_write = if bytes_copied + bytes_read as u64 <= n {
+        let bytes_to_write = if bytes_copied + bytes_read as u128 <= n {
             bytes_read
         } else {
-            n as usize - bytes_read
+            n as u128 - bytes_read
         } as usize;
         writer
             .write_all(&buf[..bytes_to_write])
             .expect("Could not write");
-        bytes_copied += bytes_to_write as u64;
+        bytes_copied += bytes_to_write as u128;
     }
     writer.flush().expect("Could not flush writer");
 }
@@ -97,40 +91,53 @@ fn cat_file(from_file_path: &PathBuf) {
     writer.flush().expect("Could not flush writer");
 }
 
-fn create_files(count: u16, target_dir: &Path) -> Vec<PathBuf> {
-    (0..count)
-        .map(|i| {
-            let file_name = format!("filler_{}", i);
-            let file_path = target_dir.join(Path::new(&file_name));
-            OpenOptions::new()
-                .write(true)
-                .create(true)
-                .read(true)
-                .truncate(true)
-                .open(&file_path)
-                .expect("Could not create filler");
-            file_path
-        })
-        .collect()
+fn config_files_already_there(
+    expected_count: usize,
+    expected_size: u128,
+    paths: &Vec<PathBuf>,
+) -> bool {
+    paths
+        .iter()
+        .filter(|p| p.exists() && p.metadata().unwrap().len() as u128 == expected_size)
+        .count()
+        == expected_count
 }
 
 fn main() {
     let cli_args = CliArgs::parse();
-
-    println!("Creating target dir...");
     let target_dir = Path::new("mfiller");
-    if Path::new("mfiller").exists() {
-        remove_dir_all(target_dir).expect("Could not remove existing target dir");
+
+    let mut files_paths = (0..cli_args.file_count)
+        .map(|i| {
+            let file_name = format!("filler_{}", i);
+            target_dir.join(Path::new(&file_name))
+        })
+        .collect();
+
+    if !config_files_already_there(cli_args.file_count.into(), cli_args.file_size, &files_paths) {
+        println!("Creating target dir...");
+        if Path::new("mfiller").exists() {
+            remove_dir_all(target_dir).expect("Could not remove existing target dir");
+        }
+        create_dir(target_dir).expect("Could not create target dir");
+
+        let formatted_bytes =
+            Byte::from_bytes(cli_args.file_size.into()).get_appropriate_unit(false);
+        println!(
+            "Creating {} files of size {}...",
+            cli_args.file_count, formatted_bytes
+        );
+        if cli_args.file_count < 1 {
+            return;
+        }
+
+        println!("Writing files content...");
+        files_paths.iter_mut().for_each(|file_path| {
+            write_content(file_path, cli_args.file_size);
+        });
+    } else {
+        println!("Using files from previous run");
     }
-    create_dir(target_dir).expect("Could not create target dir");
-
-    println!("Creating files...");
-    let mut files_paths = create_files(cli_args.file_count, &target_dir);
-
-    println!("Writing files content...");
-    files_paths.iter_mut().for_each(|file_path| {
-        write_content(file_path, cli_args.file_size);
-    });
 
     println!("Cat forever");
     loop {
